XIVY-18650 refactor: dedicated test for common model provider features
